# api/openapi.yaml
openapi: 3.1.0
info:
  title: Innerwall operator surface
  version: "1"
  summary: The public REST/JSON surface of the Innerwall control plane.
  description: |
    The operator surface is what the console and automation consume: one
    TLS-only listener beside the agent gateway, serving these endpoints under
    `/api/v1` (ADR-0021). The handlers behind them hold no domain logic; each
    calls the same domain function the control plane's command line calls, so
    the two transports cannot diverge on what is admitted or what is written
    (ADR-0007 as amended). This document describes the implementation as it
    is; where the two disagree, the disagreement is a defect in one of them,
    never something to paper over here.

    ## Authentication

    Every endpoint except `POST /session` requires a credential: the session
    cookie a login sets, or an operator token as a bearer credential. Both
    resolve to the one operator; authorization is binary. A request without
    a valid credential is a `401` problem of type `unauthenticated`. Requests
    that are not reads are refused as `403 cross-origin` when a browser marks
    them as coming from another origin.

    ## Errors

    Every error is a problem document (`application/problem+json`) with a
    stable `type` a client branches on, an HTTP `status`, a `title`, and an
    optional `detail` for people. Types are URNs under
    `urn:innerwall:problem:`; they are identifiers, not locations. Some
...

...

...

    ## Cursors

...

...

...
servers:
...
security:
  - session: []
  - bearer: []
paths:
  /session:
    post:
      operationId: createSession
      tags: [session]
      summary: Log in with the operator password.
      description: |
        The one endpoint reachable without a credential. On success a session
        cookie is set (unreadable by scripts, TLS only, not sent cross-site)
        with a fixed seven-day lifetime, and the operator is returned. A
        control plane with no password refuses with `403 no-password` and
        never offers to set one. Attempts are throttled per source address;
        beyond the limit the answer is `429 too-many-attempts` with
        `Retry-After`.
      security: []
      requestBody:
        required: true
        description: The operator password.
        content:
          application/json:
            schema:
              type: object
              required: [password]
              properties:
                password:
                  type: string
      responses:
        "200":
          description: Logged in; the cookie is set.
          headers:
            Set-Cookie:
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Me"
        "400":
          $ref: "#/components/responses/InvalidRequest"
        "401":
          description: The password is not correct (`invalid-credentials`).
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
        "403":
          description: No password has been set (`no-password`), or the request was cross-origin (`cross-origin`).
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
        "429":
          description: Too many attempts from this source in the window (`too-many-attempts`).
          headers:
            Retry-After:
              schema:
                type: integer
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
    delete:
      operationId: deleteSession
      tags: [session]
      summary: Log out.
      description: Revokes the session the request authenticated with, if it was a session, and clears the cookie either way.
      responses:
        "200":
          description: Logged out.
          content:
            application/json:
              schema:
                type: object
        "401":
          $ref: "#/components/responses/Unauthenticated"
  /me:
    get:
      operationId: getMe
      tags: [session]
      summary: The authenticated operator and the site label.
      responses:
        "200":
          description: The operator.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Me"
        "401":
          $ref: "#/components/responses/Unauthenticated"
  /flows/rollup:
    get:
      operationId: getFlowsRollup
      tags: [flows]
      summary: A grouped rollup of stored flow windows.
      description: |
        One of a closed set of named groupings over the windows in a range,
        aggregated in the store and bounded: at most `limit` groups in the
        requested order, with the totals of the whole and `truncated` set when
        more groups exist. `effective_from` and `effective_to` are the bounds
        of the windows actually covered, null when none were. Peers are what
        ingestion resolved and stored; only the current display name of a
        stored identity is added.
      parameters:
        - name: group_by
          in: query
          required: true
          description: The grouping. The set is closed; nothing else is accepted.
          schema:
            type: string
            enum: ["rule", "rule,peer", "src,dst", "dst,service"]
        - $ref: "#/components/parameters/from"
        - $ref: "#/components/parameters/to"
        - $ref: "#/components/parameters/verdict"
        - $ref: "#/components/parameters/direction"
        - name: workload
          in: query
          description: Restrict to one workload.
          schema:
            $ref: "#/components/schemas/WorkloadID"
        - $ref: "#/components/parameters/label"
        - $ref: "#/components/parameters/service"
        - name: order
          in: query
          description: The order groups are ranked in before the limit applies.
          schema:
            type: string
            enum: [connections, recent]
            default: connections
        - name: limit
          in: query
          description: Maximum groups returned; the store's default and maximum apply.
          schema:
            type: integer
            minimum: 1
            default: 200
            maximum: 1000
      responses:
        "200":
          description: The rollup.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Rollup"
        "400":
          $ref: "#/components/responses/InvalidParameter"
        "401":
          $ref: "#/components/responses/Unauthenticated"
        "404":
          $ref: "#/components/responses/NotFound"
  /flows:
    get:
      operationId: getFlows
      tags: [flows]
      summary: One page of a workload's stored flow windows.
      description: There is no unbounded form; `workload` is required.
      parameters:
        - name: workload
          in: query
          required: true
          description: The workload whose windows are listed.
          schema:
            $ref: "#/components/schemas/WorkloadID"
        - $ref: "#/components/parameters/from"
        - $ref: "#/components/parameters/to"
        - $ref: "#/components/parameters/verdict"
        - $ref: "#/components/parameters/direction"
        - name: peer
          in: query
          description: Restrict to one stored peer key (a workload id, an address group id, or a bare address).
          schema:
            type: string
        - $ref: "#/components/parameters/service"
        - $ref: "#/components/parameters/cursor"
        - name: limit
          in: query
          description: Page size; the store's default and maximum apply.
          schema:
            type: integer
            minimum: 1
            default: 100
            maximum: 500
      responses:
        "200":
          description: One page, newest window first.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FlowsPage"
        "400":
          $ref: "#/components/responses/InvalidParameter"
        "401":
          $ref: "#/components/responses/Unauthenticated"
        "404":
          $ref: "#/components/responses/NotFound"
  /workloads:
    get:
      operationId: listWorkloads
      tags: [workloads]
      summary: One page of the fleet in fleet order.
      description: |
        Fleet order is the order the fleet screen shows: the workloads needing
        attention first (degraded, offline, pending, synced), most recently
        seen first within a state.
      parameters:
        - $ref: "#/components/parameters/label"
        - name: mode
          in: query
          description: Restrict to workloads in this enforcement mode.
          schema:
            $ref: "#/components/schemas/Mode"
        - name: sync_state
          in: query
          description: Restrict to workloads in this sync state.
          schema:
            $ref: "#/components/schemas/SyncState"
        - $ref: "#/components/parameters/cursor"
        - name: limit
          in: query
          description: Page size; the read model's default and maximum apply.
          schema:
            type: integer
            minimum: 1
            default: 100
            maximum: 500
      responses:
        "200":
          description: One page of workloads.
          content:
            application/json:
              schema:
                type: object
                required: [workloads, next_cursor]
                properties:
                  workloads:
                    type: array
                    items:
                      $ref: "#/components/schemas/Workload"
                  next_cursor:
                    type: [string, "null"]
        "400":
          $ref: "#/components/responses/InvalidParameter"
        "401":
          $ref: "#/components/responses/Unauthenticated"
  /workloads/{id}:
    parameters:
      - $ref: "#/components/parameters/workloadID"
    get:
      operationId: getWorkload
      tags: [workloads]
      summary: One workload, in the shape the list uses.
      responses:
        "200":
          description: The workload.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Workload"
        "401":
          $ref: "#/components/responses/Unauthenticated"
        "404":
          $ref: "#/components/responses/NotFound"
  /workloads/{id}/rendered-policy:
    parameters:
      - $ref: "#/components/parameters/workloadID"
    get:
      operationId: getRenderedPolicy
      tags: [workloads]
      summary: The persisted rendered policy of one workload.
      description: Configuration only, no counts; the console pairs it with a rollup. A workload with no persisted policy yet is reported at version zero with no rules.
      responses:
        "200":
          description: The rendered policy.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RenderedPolicy"
        "401":
          $ref: "#/components/responses/Unauthenticated"
        "404":
          $ref: "#/components/responses/NotFound"
  /workloads/{id}/labels:
    parameters:
      - $ref: "#/components/parameters/workloadID"
    get:
      operationId: getWorkloadLabels
      tags: [workloads]
      summary: A workload's labels, as the resource a label edit is conditioned on.
      responses:
        "200":
          description: The labels and their version.
          headers:
            ETag:
              $ref: "#/components/headers/ETag"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Labels"
        "401":
          $ref: "#/components/responses/Unauthenticated"
        "404":
          $ref: "#/components/responses/NotFound"
    put:
      operationId: putWorkloadLabels
      tags: [workloads]
      summary: Replace a workload's labels.
      description: A label change moves the workload in and out of selectors, so the change renders; affected workloads' versions advance.
      parameters:
        - $ref: "#/components/parameters/ifMatch"
      requestBody:
        required: true
        description: The complete label set that replaces the current one.
        content:
          application/json:
            schema:
              type: object
              required: [labels]
              properties:
                labels:
                  $ref: "#/components/schemas/LabelMap"
      responses:
        "200":
          description: The labels as persisted, with their new version.
          headers:
            ETag:
              $ref: "#/components/headers/ETag"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Labels"
        "400":
          $ref: "#/components/responses/Validation"
        "401":
          $ref: "#/components/responses/Unauthenticated"
        "404":
          $ref: "#/components/responses/NotFound"
        "412":
          $ref: "#/components/responses/PreconditionFailed"
        "428":
          $ref: "#/components/responses/PreconditionRequired"
  /rulesets:
    get:
      operationId: listRulesets
      tags: [policy]
      summary: Every ruleset, and the version of the state a policy editor authors against.
      description: |
        `state_version` is a digest of everything a render reads: each
        workload's labels, addresses, and mode, and each authored object's
        last write. A dry run reports the version it computed against, so an
        editor that started from this value can tell when the state moved.
      responses:
        "200":
          description: The rulesets.
          content:
            application/json:
              schema:
                type: object
                required: [rulesets, state_version]
                properties:
                  rulesets:
                    type: array
                    items:
                      $ref: "#/components/schemas/Ruleset"
                  state_version:
                    $ref: "#/components/schemas/StateVersion"
        "401":
          $ref: "#/components/responses/Unauthenticated"
    post:
      operationId: createRuleset
      tags: [policy]
      summary: Create a ruleset with its rules.
      description: The whole ruleset is admitted before anything is persisted, and every workload is rendered after.
      requestBody:
        $ref: "#/components/requestBodies/Ruleset"
      responses:
        "201":
          description: The ruleset as persisted.
          headers:
            ETag:
              $ref: "#/components/headers/ETag"
            Location:
              $ref: "#/components/headers/Location"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Ruleset"
        "400":
          $ref: "#/components/responses/Validation"
        "401":
          $ref: "#/components/responses/Unauthenticated"
        "409":
          $ref: "#/components/responses/DuplicateName"
  /rulesets/{id}:
    parameters:
      - $ref: "#/components/parameters/id"
    get:
      operationId: getRuleset
      tags: [policy]
      summary: One ruleset with its rules.
      responses:
        "200":
          description: The ruleset.
          headers:
            ETag:
              $ref: "#/components/headers/ETag"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Ruleset"
        "401":
          $ref: "#/components/responses/Unauthenticated"
        "404":
          $ref: "#/components/responses/NotFound"
    put:
      operationId: putRuleset
      tags: [policy]
      summary: Replace a ruleset and all of its rules.
      description: A rule that keeps its id keeps its `created_at`, and its `updated_at` and version move only when the rule itself changed; a rule without an id is new.
      parameters:
        - $ref: "#/components/parameters/ifMatch"
      requestBody:
        $ref: "#/components/requestBodies/Ruleset"
      responses:
        "200":
          description: The ruleset as persisted.
          headers:
            ETag:
              $ref: "#/components/headers/ETag"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Ruleset"
        "400":
          $ref: "#/components/responses/Validation"
        "401":
          $ref: "#/components/responses/Unauthenticated"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/DuplicateName"
        "412":
          $ref: "#/components/responses/PreconditionFailed"
        "428":
          $ref: "#/components/responses/PreconditionRequired"
    delete:
      operationId: deleteRuleset
      tags: [policy]
      summary: Delete a ruleset and its rules.
      parameters:
        - $ref: "#/components/parameters/ifMatch"
      responses:
        "204":
          description: Deleted.
        "401":
          $ref: "#/components/responses/Unauthenticated"
        "404":
          $ref: "#/components/responses/NotFound"
        "412":
          $ref: "#/components/responses/PreconditionFailed"
        "428":
          $ref: "#/components/responses/PreconditionRequired"
  /rulesets/{id}/rules:
    parameters:
      - $ref: "#/components/parameters/id"
    post:
      operationId: createRule
      tags: [policy]
      summary: Add one rule to a ruleset.
      description: The ruleset is written as a unit, conditioned on the version it was read at inside the control plane, so a concurrent edit is refused rather than overwritten. Findings are relative to the rule.
      requestBody:
        $ref: "#/components/requestBodies/Rule"
      responses:
        "201":
          description: The rule as persisted.
          headers:
            ETag:
              $ref: "#/components/headers/ETag"
            Location:
              $ref: "#/components/headers/Location"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Rule"
        "400":
          $ref: "#/components/responses/Validation"
        "401":
          $ref: "#/components/responses/Unauthenticated"
        "404":
          $ref: "#/components/responses/NotFound"
        "412":
          $ref: "#/components/responses/PreconditionFailed"
  /rulesets/{id}/rules/{rule_id}:
    parameters:
      - $ref: "#/components/parameters/id"
      - name: rule_id
        in: path
        required: true
        description: The rule's id.
        schema:
          type: string
          format: uuid
    get:
      operationId: getRule
      tags: [policy]
      summary: One rule.
      responses:
        "200":
          description: The rule.
          headers:
            ETag:
              $ref: "#/components/headers/ETag"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Rule"
        "401":
          $ref: "#/components/responses/Unauthenticated"
        "404":
          $ref: "#/components/responses/NotFound"
    put:
      operationId: putRule
      tags: [policy]
      summary: Replace one rule, conditioned on the rule's own version.
      parameters:
        - $ref: "#/components/parameters/ifMatch"
      requestBody:
        $ref: "#/components/requestBodies/Rule"
      responses:
        "200":
          description: The rule as persisted.
          headers:
            ETag:
              $ref: "#/components/headers/ETag"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Rule"
        "400":
          $ref: "#/components/responses/Validation"
        "401":
          $ref: "#/components/responses/Unauthenticated"
        "404":
          $ref: "#/components/responses/NotFound"
        "412":
          $ref: "#/components/responses/PreconditionFailed"
        "428":
          $ref: "#/components/responses/PreconditionRequired"
    delete:
      operationId: deleteRule
      tags: [policy]
      summary: Remove one rule, conditioned on the rule's own version.
      parameters:
        - $ref: "#/components/parameters/ifMatch"
      responses:
        "204":
          description: Deleted.
        "401":
          $ref: "#/components/responses/Unauthenticated"
        "404":
          $ref: "#/components/responses/NotFound"
        "412":
          $ref: "#/components/responses/PreconditionFailed"
        "428":
          $ref: "#/components/responses/PreconditionRequired"
...
  /provisioning-tokens/{id}:
    parameters:
      - $ref: "#/components/parameters/id"
    delete:
      operationId: revokeProvisioningToken
      tags: [tokens]
      summary: Revoke a provisioning token.
      description: Revocation, not deletion; the token stays listed as revoked. Future enrollments with it are refused; workloads it already enrolled are untouched.
      responses:
        "204":
          description: Revoked.
        "401":
          $ref: "#/components/responses/Unauthenticated"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/AlreadyRevoked"
...
components:
  securitySchemes:
    session:
      type: apiKey
      in: cookie
      name: innerwall_session
      description: The session cookie a login sets.
    bearer:
      type: http
      scheme: bearer
      description: An operator token (`iwo_...`) as a bearer credential.
...